+++ src/main/java/src/backend/controller/ExpenseController.java
@@ -73,6 +73,23 @@ public class ExpenseController extends ControllerTemplate {
         return result;
     }
 
+    public List<Expense> getFavoriteExpenses() {
+
+        List<Expense> result = new ArrayList<>();
+
+        try {
+            result = expenseService.getFavoriteExpenses();
+
+            Logger.info("Successfully retrieved favorite expenses.");
+        } catch (SQLException sqle) {
+            Logger.error(SQL_ERROR_TEXT, sqle.getMessage());
+        } catch (Exception e) {
+            Logger.error(GENERAL_ERROR_TEXT, e.getMessage());
+        }
+
+        return result;
+    }
+
     public UUID updateExpense(Expense expense) {
 
         UUID result = null;